+++ Backend/FileSystem/Rmusr.go
@@ -48,7 +48,7 @@ func Rmusr(name string, currentUser *structs.Sesion, activeSession *bool, w http
 	file.Seek(ToInt(sp.S_inode_start[:]), os.SEEK_SET)
 	ReadInode(&root, file)
 
-	pointerOfFile := 0
+	pointerOfFile := int64(0)
 
 	root = SearchFile(file, root, SplithPath("users.txt"), ToInt(sp.S_inode_start[:]), ToInt(sp.S_block_start[:]), &pointerOfFile)
