+++ Backend/Structures/Block.go
@@ -2,7 +2,7 @@ package structures
 
 type Content struct {
 	B_name  [12]byte
-	B_inodo int64
+	B_inodo int32
 }
 
 type Dirblock struct {
